+++ app/models/supply_teachers/journey/fta_to_perm_contract_start.rb
@@ -3,7 +3,7 @@ module SupplyTeachers
     DATE_ATTIBUTES = %i[contract_start_date].freeze
 
     include Steppable
-    include DateValidator
+    include DateValidations
     include ContractStartable
 
     def next_step_class